Major update to layout and design of about us pages. The tab selection colors have been changed. Moreover, the Team part which was previously included in the product subpage has now it's own tab. Therefore a new page was created and appriopiate routes have been added. Dividers have been removed to be coherent with titles.
